Implemented scalable form structure to Enquirt form component.
