+++ src/components/ExternalFormView.tsx
@@ -1,0 +1,6 @@
+const ExternalFormView = () => {
+  return (
+    <div>ExternalFormView</div>
+  )
+};
+export  default ExternalFormView;
